requirements.txt
keras==2.10.0
matplotlib==3.5.2
numba==0.53.0
numpy==1.23.3
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.3.0
plotly==5.11.0
rasterfairy==1.0.6
rasterfairy_py3==1.0.5
scikit_learn==1.2.0
seaborn==0.11.2
tensorflow==2.10.0
tflite_runtime==2.5.0.post1
tqdm==4.64.1
umap==0.1.1
umap_learn==0.5.3
